refactored follder structure
