+++ Decsys/ClientApp/package.json
@@ -6,9 +6,11 @@
     "@smooth-ui/core-sc": "^8.1.0",
     "react": "^16.7.0",
     "react-dom": "^16.7.0",
+    "react-redux": "^6.0.0",
     "react-router-dom": "^4.3.1",
     "react-scripts": "2.1.3",
     "react-tabulator": "^0.7.11",
+    "redux": "^4.0.1",
     "styled-components": "^4.1.3",
     "styled-icons": "^6.6.0"
   },
